[fix] correct oauth gg
