+++ libs/api/auth/data-access/src/lib/auth-service.module.ts
@@ -1,11 +1,21 @@
 import { Global, Module } from '@nestjs/common';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+import { JwtModule } from '@nestjs/jwt';
 import { AuthService } from './auth.service';
-import { JwtService } from '@nestjs/jwt';
 
 @Global()
 @Module({
   controllers: [],
-  providers: [AuthService, JwtService],
+  providers: [AuthService],
+  imports: [
+    JwtModule.registerAsync({
+      imports: [ConfigModule],
+      useFactory: async (configService: ConfigService) => ({
+        secret: configService.get('JWT_SECRET'),
+      }),
+      inject: [ConfigService],
+    }),
+  ],
   exports: [AuthService],
 })
 export class AuthServiceModule {}
