+++ libs/api/auth/feature/src/lib/auth.controller.ts
@@ -1,7 +1,11 @@
-import { Controller, Post, Body, Get } from "@nestjs/common";
+import { Controller, Post, Body, Get, Req, Res, UseGuards } from "@nestjs/common";
+import { GoogleOAuthGuard } from "./guards";
+import { Response } from 'express';
+import { AuthService } from '@yw/api/auth/data-access';
 
 @Controller("auth")
 export class AuthController {
+  constructor(private authService: AuthService) {}
   
   @Post("sign-up")
   signUp(@Body() userData: any) {
@@ -18,5 +22,23 @@ export class AuthController {
   @Get("")
   refreshToken() {
     return true;
+  }
+
+  @UseGuards(GoogleOAuthGuard)
+  @Get('google')
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
+  async googleAuth() {}
+
+  @UseGuards(GoogleOAuthGuard)
+  @Get('google-auth-redirect')
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  async googleAuthRedirect(@Req() req: any, @Res() res: Response) {
+    const { encodedUser } = await this.authService.signInWithGoogle(
+      req.user,
+      res
+    );
+    return res.redirect(
+      `${process.env['GOOGLE_REDIRECT_URL_CLIENT_ANGULAR']}?jwtUser=${encodedUser}`
+    );
   }
 }
